pyproject.toml
[project]
name = "mcpservers"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio>=4.9.0",
    "gradio>=5.23.1",
    "openai-agents>=0.0.7",
    "planningmcp",
    "python-dotenv>=1.1.0",
    "templatemcp",
    "typer>=0.15.2",
]

[project.scripts]
mcpservers = "mcpservers.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.workspace]
members = ["servers/templatemcp", "servers/planningmcp"]

[tool.uv.sources]
templatemcp = { workspace = true }
planningmcp = { workspace = true }

[tool.ruff]
exclude = ["build"]
line-length = 120

[tool.ruff.lint]
select = [
    "B",   # flake8-bugbear
    "C",   # flake8-comprehensions
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "SIM", # flake8-simplify
    "UP",  # pyupgrade
    "W",   # pycodestyle warnings
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
ignore_missing_imports = true

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pip>=25.0.1",
    "types-aiofiles>=24.1.0.20250326",
    "types-pytz>=2025.2.0.20250326",
    "types-pyyaml>=6.0.12.20250326",
    "types-simplejson>=3.20.0.20250326",
    "types-ujson>=5.10.0.20250326",
]
